Implemented clearConsole method, displayFullPost method and modified the constructor
Constructor now accepts only one string, the creator's name(later this String will be taken from the database) and asks for the text. DisplayFullPost method has the role of the previous dummy mai, meaning that it sums up the basic operations of the Post class. ClearConsole checks for the os and clear the console to make the programm better-looking